requirements.txt
matplotlib==3.5.0
numpy==1.21.2
pandas==1.4.0
Pillow==9.0.1
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.8.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
transformers==4.16.2
